Realized that most operations should flow through mod-users-bl
I had assumed that mod-users endpoints were exposed to the UI, but now realize that they don't
appear to be. For example, I wanted to try resetting a password for a newly created user, but
while debugging that I noticed that the UI is sending everything through mod-users-bl endpoints.
So it may be that mod-users in fact isn't a good candidate for integration tests. In other words
mod-login's routes appear to be called exclusively through mod-users-bl, which does more than
simply pass requests through mod-users.