Morphing mod-login into mod-users-bl
Based on discussions with the CP team and contributors responsible for mod-users-bl, I'm
morphing the mod-login project into a mod-users-bl project. It was determined that mod-login
because it doesn't talk to the top layer of our code directly wouldn't benefit from its own
mod-users project in FAT. My goal is to write a few mod-users-bl tests to get things rolling
for mod-users-bl. Unlike mod-login, mod-users-bl is called extensively by the UI. Because of
this its routes will benefit the most from integration-style testing.