FAT-933: Implement karate tests for mod-users-bl and mod-login (#244)
Resolves FAT-933, FAT-929, and FAT-931.

All three tickets were merged into one to speed up implementation and code review.

Based on discussions with the CP team and contributors responsible for mod-users-bl, I
morphed the mod-login project I started with into a mod-users-bl project. It was determined
that mod-login--because it doesn't talk to the top layer of our code directly--wouldn't benefit
from its own mod-login project in FAT. My goal is to write a few mod-users-bl tests to get things
rolling for mod-users-bl. Unlike mod-login, mod-users-bl is called extensively by the UI. Because of
this, its routes will benefit the most from integration-style testing.