fix(holdings-items): remove holdingsItems field (#1153)
Previously, during setup.feature in mod-quick-marc on the "Create MARC-HOLDINGS record" scenario tests failed due to the features introduced in the MODINVSTOR-1094. Also, the similar situation appeared in the create-test-data.feature in mod-search on the "Create inventory holdings" scenario. This commit solves these errors by deleting holdingsItems and bareHoldingsItems fields.

- Remove holdingsItems and bareHoldingsItems fields from the holdings.json file in mod-quick-marc and mod-search modules.

Closes FAT-9832